+++ src/ila/event/minecraft/world/actor/ArmorStandSwapItemEvent.h
@@ -2,7 +2,7 @@
 #include "ila/base/Macro.h"
 #include <ll/api/event/Cancellable.h>
 #include <ll/api/event/player/PlayerEvent.h>
-#include <mc/deps/shared_types/EquipmentSlot.h>
+#include <mc/deps/shared_types/legacy/item/EquipmentSlot.h>
 
 namespace ila::mc::inline actor
 {