+++ src/ila/event/minecraft/world/actor/player/PlayerAttackBlockEvent.h
@@ -1,20 +1,23 @@
 #pragma once
 #include "ila/base/Macro.h"
 #include <ll/api/event/Cancellable.h>
-#include <ll/api/event/player/PlayerEvent.h>
+#include <ll/api/event/player/ServerPlayerEvent.h>
+#include <mc/common/FacingID.h>
 #include <mc/world/level/BlockPos.h>
 
 namespace ila::mc::inline player
 {
-class PlayerAttackBlockBeforeEvent final : public ll::event::Cancellable<ll::event::player::PlayerEvent>
+class PlayerAttackBlockBeforeEvent final : public ll::event::Cancellable<ll::event::player::ServerPlayerEvent>
 {
 protected:
     BlockPos& mPos;
+    FacingID& mFace;
 
 public:
-    constexpr explicit PlayerAttackBlockBeforeEvent(Player& player, BlockPos& pos)
+    constexpr explicit PlayerAttackBlockBeforeEvent(ServerPlayer& player, BlockPos& pos, FacingID& face)
         : Cancellable(player)
         , mPos(pos)
+        , mFace(face)
     {
     }
 
@@ -22,22 +25,30 @@ public:
     ILAPI void deserialize(CompoundTag const& nbt) override;
 
     ILNDAPI BlockPos& getPos() const;
+    ILNDAPI FacingID& getFace() const;
 };
 
-class PlayerAttackBlockAfterEvent final : public ll::event::player::PlayerEvent
+class PlayerAttackBlockAfterEvent final : public ll::event::player::ServerPlayerEvent
 {
 protected:
     BlockPos const& mPos;
+    FacingID const& mFace;
 
 public:
-    constexpr explicit PlayerAttackBlockAfterEvent(Player& player, BlockPos const& pos)
-        : PlayerEvent(player)
+    constexpr explicit PlayerAttackBlockAfterEvent(
+        ServerPlayer&   player,
+        BlockPos const& pos,
+        FacingID const& face
+    )
+        : ServerPlayerEvent(player)
         , mPos(pos)
+        , mFace(face)
     {
     }
 
     ILAPI void serialize(CompoundTag& nbt) const override;
 
     ILNDAPI BlockPos const& getPos() const;
+    ILNDAPI FacingID const& getFace() const;
 };
 } // namespace ila::mc::inline player