+++ src/ila/event/minecraft/world/actor/player/PlayerOpenContainerEvent.h
@@ -1,7 +1,7 @@
 #include "ila/base/Macro.h"
 #include <ll/api/event/Cancellable.h>
 #include <ll/api/event/player/ServerPlayerEvent.h>
-#include <mc/common/ActorUniqueID.h>
+#include <mc/legacy/ActorUniqueID.h>
 
 namespace ila::mc::inline player
 {
@@ -9,18 +9,18 @@ class PlayerOpenContainerBeforeEvent final
     : public ll::event::Cancellable<ll::event::player::ServerPlayerEvent>
 {
 protected:
-    BlockPos&      mPos;
-    ContainerID&   mContainerId;
-    ContainerType& mContainerType;
-    ActorUniqueID& mContainerActorId;
+    BlockPos&                           mPos;
+    ContainerID&                        mContainerId;
+    SharedTypes::Legacy::ContainerType& mContainerType;
+    ActorUniqueID&                      mContainerActorId;
 
 public:
     constexpr explicit PlayerOpenContainerBeforeEvent(
-        ServerPlayer&  player,
-        BlockPos&      pos,
-        ContainerID&   containerId,
-        ContainerType& containerType,
-        ActorUniqueID& containerActorId
+        ServerPlayer&                       player,
+        BlockPos&                           pos,
+        ContainerID&                        containerId,
+        SharedTypes::Legacy::ContainerType& containerType,
+        ActorUniqueID&                      containerActorId
     )
         : Cancellable(player)
         , mPos(pos)
@@ -33,27 +33,27 @@ public:
     ILAPI void serialize(CompoundTag& nbt) const override;
     ILAPI void deserialize(CompoundTag const& nbt) override;
 
-    ILNDAPI BlockPos&      getContainerBlockPos() const;
-    ILNDAPI ContainerID&   getContainerId() const;
-    ILNDAPI ContainerType& getContainerType() const;
-    ILNDAPI ActorUniqueID& getContainerActorId() const;
+    ILNDAPI BlockPos&    getContainerBlockPos() const;
+    ILNDAPI ContainerID& getContainerId() const;
+    ILNDAPI SharedTypes::Legacy::ContainerType& getContainerType() const;
+    ILNDAPI ActorUniqueID&                      getContainerActorId() const;
 };
 
 class PlayerOpenContainerAfterEvent final : public ll::event::player::ServerPlayerEvent
 {
 protected:
-    BlockPos const&      mPos;
-    ContainerID const&   mContainerId;
-    ContainerType const& mContainerType;
-    ActorUniqueID const& mContainerActorId;
+    BlockPos const&                           mPos;
+    ContainerID const&                        mContainerId;
+    SharedTypes::Legacy::ContainerType const& mContainerType;
+    ActorUniqueID const&                      mContainerActorId;
 
 public:
     constexpr explicit PlayerOpenContainerAfterEvent(
-        ServerPlayer&        player,
-        BlockPos const&      pos,
-        ContainerID const&   containerId,
-        ContainerType const& containerType,
-        ActorUniqueID const& containerActorId
+        ServerPlayer&                             player,
+        BlockPos const&                           pos,
+        ContainerID const&                        containerId,
+        SharedTypes::Legacy::ContainerType const& containerType,
+        ActorUniqueID const&                      containerActorId
     )
         : ServerPlayerEvent(player)
         , mPos(pos)
@@ -65,9 +65,9 @@ public:
 
     ILAPI void serialize(CompoundTag& nbt) const override;
 
-    ILNDAPI BlockPos const&      getContainerBlockPos() const;
-    ILNDAPI ContainerID const&   getContainerId() const;
-    ILNDAPI ContainerType const& getContainerType() const;
-    ILNDAPI ActorUniqueID const& getContainerActorId() const;
+    ILNDAPI BlockPos const&    getContainerBlockPos() const;
+    ILNDAPI ContainerID const& getContainerId() const;
+    ILNDAPI SharedTypes::Legacy::ContainerType const& getContainerType() const;
+    ILNDAPI ActorUniqueID const&                      getContainerActorId() const;
 };
 } // namespace ila::mc::inline player