+++ src/ila/include_all.h
@@ -3,6 +3,7 @@
 #include "ila/event/legacyMoney/MoneyChangeEvent.h"
 #include "ila/event/leviAntiCheat/PlayerBanWaveEvent.h"
 #include "ila/event/leviAntiCheat/PlayerCheatEvent.h"
+#include "ila/event/leviAntiCheat/SusClientEvent.h"
 #include "ila/event/minecraft/server/ClientLoginEvent.h"
 #include "ila/event/minecraft/server/RegisterCmdEvent.h"
 #include "ila/event/minecraft/server/SendPacketEvent.h"